itemfield.cpp: gcc6/C++14 compatibility
Recent gcc and/or C++14 changed the definition of abs() such that
applying it to a 16bit integer resulted in a float, leading to:
itemfield.cpp:1491:27: error: invalid operands of types
'__gnu_cxx::__enable_if<true, double>::__type {aka double}' and 'const int' to
binary 'operator%'
(uInt16)(abs(moffs) % MinsPerHour)

+++ src/sysync/itemfield.cpp
@@ -1487,8 +1487,8 @@ void TTimestampField::getAsRFC822date(string &aRFC822String, timecontext_t aTarg
       aRFC822String,
       " %c%02hd%02hd",
       moffs>=0 ? '+' : '-',
-      (uInt16)(abs(moffs) / MinsPerHour),
-      (uInt16)(abs(moffs) % MinsPerHour)
+      (uInt16)(abs((int)moffs) / MinsPerHour),
+      (uInt16)(abs((int)moffs) % MinsPerHour)
     );
   }
 } // TTimestampField::getAsRFC822date
